Refactor: 모임일정 관련 비슷한 기능들 한개의 모듈로 통합
