+++ Namo_SwiftUI/Projects/Feature/Gathering/Example/Sources/FeatureGatheringScheduleExampleTemplate.swift
