+++ Namo_SwiftUI/Projects/Feature/Gathering/Sources/FeatureGatheringScheduleSourcesTemplate.swift
