+++ Namo_SwiftUI/Projects/Feature/GroupTab/Interface/Sources/FeatureGroupListInterfaceTemplate.swift
