+++ Namo_SwiftUI/Projects/Feature/Sources/FeatureExport.swift
@@ -11,12 +11,9 @@
 @_exported import FeatureOnboarding
 @_exported import FeatureCalendar
 @_exported import FeatureArchive
-@_exported import FeatureGroupList
-@_exported import FeatureGatheringListInterface
-@_exported import FeatureGatheringList
-@_exported import FeatureGatheringScheduleInterface
-@_exported import FeatureGatheringSchedule
-@_exported import FeatureKakaoMapInterface
+@_exported import FeatureGroupTab
+@_exported import FeatureGathering
+@_exported import FeatureGatheringInterface
 @_exported import FeatureFriendInviteInterface
 @_exported import FeatureLocationSearchInterface
 @_exported import FeatureLocationSearch
